examples/flask/requirements.txt
Flask==2.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
python-dotenv